Now able to read the XDR language.
I'm now able to read the XDR language, to make sense of it. Able to
consume the entire NULL procedure, no extra bytes at the end of the
buffer. Should be able to start writing parsers for the compound
procedure; therefore should be able to go by the NFS specification to
implement NFS.

The very next step is to go back into Packet and see if I'm able to fix
the Packet language so that I'm able define these two packets, the NULL
call and response.